Bug 78 bool (#101)
* updates readme

* defaults boolean vals to false (so checkbox works)

* recursive method to replace boolean widgets, need to fix

* -

* manual ui schema in progress

* ephys ui schema

* fix ephys ui schema

* endline

* small fix

* Bug 78 bool widget (#103)

* feat: adds rjsf material ui package

* feat: adds custom checkbox for booleans

---------


* removes override

---------
