+++ src/components/RenderForm.js
@@ -1,8 +1,8 @@
-import React from 'react';
 import Form from '@rjsf/core';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import validator from '@rjsf/validator-ajv8';
-import {widgets, timeUiSchema} from '../custom-ui/TimeUISchema';
+import { widgets } from '../custom-ui/CustomWidgets';
+import { uiSchema } from '../custom-ui/CustomUISchema';
 
 export default function RenderForm (props) {
   /*
@@ -37,18 +37,18 @@ export default function RenderForm (props) {
    writer.close();
   }
 
-
   if(schema){
       return (
         schema && <Form schema={schema}
         formData={formData}
         validator={validator}
-        uiSchema={timeUiSchema}
+        uiSchema={uiSchema}
         widgets={widgets}
         onSubmit={saveFilePicker} 
         noHtml5Validate >
         </Form>
       );
+
  } else {
   return(<div> Please select a schema from the dropdown above. </div>)
  }
