+++ src/custom-ui/CustomUISchema.js
@@ -1,0 +1,6 @@
+export const uiSchema = {
+    "ui:options": {
+        timeFormat: "hh:mm:ss",
+        "ui:placeholder": "00:00:00"
+    }
+};
