+++ src/custom-ui/CustomWidgets.js
@@ -1,0 +1,38 @@
+import Datetime from 'react-datetime';
+import 'react-datetime/css/react-datetime.css';
+import moment from 'moment';
+import {default as RadioWidget} from "@rjsf/material-ui/lib/RadioWidget/RadioWidget";
+import {default as CheckboxWidget} from "@rjsf/material-ui/lib/CheckboxWidget/CheckboxWidget";
+
+const CustomTimeWidget = (props) => {
+    const onChange = (selectedDate) => {
+        const formattedTime = moment(selectedDate).format('HH:mm:ss');
+        props.onChange(formattedTime);
+      };
+  return (
+    <Datetime
+      dateFormat={false}
+      timeFormat="HH:mm:ss"
+      value={props.value ? moment(props.value, 'HH:mm:ss') : undefined}
+      onChange={onChange}
+    />
+  );
+};
+
+/**
+ * Uses RadioButtons for booleans so
+ * users understand undefined default
+ */
+const CustomCheckboxWidget = (props) => {
+    if (props.schema.type === "boolean"){
+        return(
+            RadioWidget(props)
+       );
+    } else {
+       return (
+          CheckboxWidget(props)
+       );
+    };
+};
+
+export const widgets = {checkbox: CustomCheckboxWidget, time: CustomTimeWidget};
